* move midi to communication folder
